fix: Múltiples correcciones en módulo de compras
1. Eliminar botón "Recibir Mercancía" de detalles de compra
   - El botón estaba de más en la vista de detalles

2. Corregir soporte para cantidades decimales (80.71 lb)
   - Cambiar columnas quantity y quantity_received a DECIMAL(12,3)
   - Usar parseFloat() en lugar de parseInt() al leer cantidades
   - Agregar migración automática en ambos endpoints de compras

3. Permitir edición de compras por admin/manager
   - Admin, Super Admin y Market Manager pueden editar compras confirmadas
   - Solo no se pueden editar compras ya recibidas o canceladas
   - Comerciales solo pueden editar borradores

+++ src/app/api/market/purchases/route.ts
@@ -308,8 +308,8 @@ export async function GET(request: NextRequest) {
           confirmedByName: row.confirmed_by_name,
           receivedByName: row.received_by_name,
           lineCount: parseInt(row.line_count) || 0,
-          totalItems: parseInt(row.total_items) || 0,
-          totalReceived: parseInt(row.total_received) || 0
+          totalItems: parseFloat(row.total_items) || 0,
+          totalReceived: parseFloat(row.total_received) || 0
         })),
         stats: {
           draft: parseInt(stats.draft_count) || 0,
